0424/최민승/업로드관련모든작업끝/

+++ src/main/java/com/campEZ/Project0/uploadfile/dao/UploadFileDAO.java
@@ -23,5 +23,5 @@ public interface UploadFileDAO {
 
   int deleteFileByUploadFileId(int uploadfileId);
 
-  int deleteFileByCodeWithRid(AttachFileType attachFileType, Long rid);
+  int deleteFileByCodeWithRid(AttachFileType attachFileType, int rid);
 }
